remove dead link to heroku
I stopped the Heroku service because they removed free hosting. I was using the basic paid service when I needed it, but didn't want to pay for the hosting at times when it was rarely used. It was all the easier to drop Heroku since they forced the use of MFA and I nearly got locked out. Plus being taken over by Salesforce wasn't a good look.